backend/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
Flask_Cors==4.0.0
mysql_connector_repackaged==0.3.1
Pillow==9.3.0
Pillow==9.4.0
Pillow==10.0.1
Pillow==10.0.1
PyJWT==1.4.2
PyJWT==2.4.0
PyJWT==2.8.0
PyJWT==2.8.0
torch==2.0.0+cu117
torch==2.0.0
torch==1.12.1
torch==2.0.1
torchvision==0.15.1+cu117
torchvision==0.15.2
